nfs: domain uuid directory must be specified as destination
Previously no check was done on -n parameter
for ensuring that the destination was an export domain
uuid directory.
Now the -n parameter is validated checking for
the export domain uuid directory as destination.
The output of --help and the man page have been updated
to explicitly refer to the domain uuid directory.

Additionally, the path to the log direcotry has been
updated in the man page.

Change-Id: Ic3e2f1ab6a6788a0ce80130c32171f69e7562c56
Bug-Url: https://bugzilla.redhat.com/950634
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>