UX: remove alarming warning
After upgrading from 3.1 users was warned about a legacy config
still present in conf files if the user forgot to update .conf file
with the one provided by the package as .rpmnew.
The tool is able to use the legacy configuration so no need to
alarm the user about that.

Change-Id: I2c2d666b57721ec5f3fbb7dc70319a5745521908
Bug-Url: https://bugzilla.redhat.com/1031605
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>