Renamed engine-image-uploader to ovirt-image-uploader
The package should have the same name of the application.
In this case, the package name seems better than the application
name, so I've renamed the application instead of the package.

engine-image-uploader and a placeholder for its man page are kept
for backward compatibility until all the documentation will be
updated.

Change-Id: Ifd6524d4655b216b9746c4ffa68121484771f9ed
Signed-off-by: Sandro Bonazzola <sbonazzo@redhat.com>